norm heredoc, cmd_line_struct
